2.9.1

+++ Plain Craft Launcher 2/Pages/PageLaunch/PageLaunchLeft.xaml.vb
@@ -53,10 +53,10 @@
             '自动整合包安装
             If PackInstallPath IsNot Nothing Then
                 Try
-                    Dim InstallLoader = ModpackInstall(PackInstallPath, GetFolderNameFromPath(Path))
+                    Dim InstallLoader = ModpackInstall(PackInstallPath)
                     Log("[Launch] 自动安装整合包已开始：" & PackInstallPath)
                     InstallLoader.WaitForExit()
-                    If InstallLoader.State = LoadState.Finished Then
+                    If InstallLoader.State = LoadState.Finished AndAlso File.Exists(PackInstallPath) Then
                         Log("[Launch] 自动安装整合包成功，删除安装包：" & PackInstallPath)
                         File.Delete(PackInstallPath)
                     End If
